Refactor plugins to have a shared base plugin
